Applied the proposed changes - To run the project, use the "go run . -mode=sequential"  or "go run . -mode=concurrent". By default, the project uses mode sequential.
