--- backend/go.mod
@@ -1,5 +1,0 @@
-module xkcd/comics
-
-go 1.22.1
-
-require github.com/mattn/go-sqlite3 v1.14.22 // indirect
