--- backend/model/comic.go
@@ -1,13 +1,0 @@
-package model
-
-type Comic struct {
-    Num        int    `json:"num"`
-    SafeTitle  string `json:"safe_title"`
-    Img        string `json:"img"`
-    Title      string `json:"title"`
-    Transcript string `json:"transcript"`
-}
-
-type LatestComic struct {
-    Num int `json:"num"`
-}
