+++ go.mod
@@ -1,0 +1,3 @@
+module xkcdcomics
+
+go 1.22.1
